moved reading and handling user input into debugger_commands.c
